Updated module augmentation for highlight package.
